* branches/rudder/2.4:
  fixes #2560 Cpu speed is now an Integer in ldap, search on that field are now working
  Fixes #2565 : when a node has been removed, and a new inventory arrives for it, it comes back into pending

+++ inventory-fusion/src/main/scala/com/normation/inventory/provisioning/fusion/RudderParsingRules.scala
@@ -144,7 +144,7 @@ object RudderCpuParsing extends FusionReportParsingExtension with Loggable {
         case Some(name) =>
           val cpu = Processor(
               name = name
-            , speed = optText(p\"FREQUENCY").map(_.toFloat)
+            , speed = optText(p\"FREQUENCY").map(_.toFloat.toInt)
             , model = optText(p\"MODEL")
             , family = optText(p\"FAMILY")
             , manufacturer = optText(p\"VENDOR") map { Manufacturer(_) }
